refactor: rename helpers.ts to OctokitProvider.ts and add docstrings
