+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/github/githubWebhook.ts
@@ -15,7 +15,7 @@
  */
 import { ScmIntegrationRegistry } from '@backstage/integration';
 import { createTemplateAction } from '../../createTemplateAction';
-import { OctokitProvider } from './helpers';
+import { OctokitProvider } from './OctokitProvider';
 
 type ContentType = 'form' | 'json';
